task1/requirements.txt
Appium-Python-Client==2.0
PyYAML~=6.0.2
uiautomator2~=3.2.5
selenium~=4.0.0
numpy~=2.1.1
pytest~=8.3.3
requests~=2.32.3